+++ src/components/CategoryFilter.js
@@ -12,8 +12,8 @@ export const CATEGORIES = [
 ];
 
 const CategoryFilter = ({ handleFilter }) => (
-  <label htmlFor="filterCategories" className="form-label">
-    Filter category
+  <label htmlFor="filterCategories" className="form-label d-flex">
+    CATEGORIES&nbsp;
     <select name="filterCategories" onChange={handleFilter} id="filterCategories" className="form-select">
       <option value="All">All</option>
       {
