+++ src/containers/BooksList.js
@@ -2,34 +2,17 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import Book from '../components/Book';
-import CategoryFilter from '../components/CategoryFilter';
-import {
-  CHANGE_FILTER,
-  CHANGE_VISIBILITY,
-  deleteBook,
-} from '../actions';
+import { deleteBook } from '../actions';
 
-const BookList = ({
-  books,
-  delete: handleRemoveBook,
-  handleChangeFilter,
-  handleVisibility,
-}) => {
-  const handleFilter = (e) => {
-    handleChangeFilter(e.target.value);
-    handleVisibility(e.target.value);
-  };
-  return (
-    <div>
-      <CategoryFilter handleFilter={handleFilter} />
-      {
-        books.map(
-          (book) => <Book book={book} key={book.id} delete={handleRemoveBook} />,
-        )
-      }
-    </div>
-  );
-};
+const BookList = ({ books, delete: handleRemoveBook }) => (
+  <div>
+    {
+      books.map(
+        (book) => <Book book={book} key={book.id} delete={handleRemoveBook} />,
+      )
+    }
+  </div>
+);
 
 BookList.propTypes = {
   books: PropTypes.arrayOf(
@@ -40,16 +23,10 @@ BookList.propTypes = {
     }),
   ).isRequired,
   delete: PropTypes.func.isRequired,
-  handleVisibility: PropTypes.func.isRequired,
-  handleChangeFilter: PropTypes.func.isRequired,
 };
 
 const mapStateToProps = (state) => ({ books: state.books });
 
-const mapDispatchToProps = (dispatch) => ({
-  delete: (book) => dispatch(deleteBook(book)),
-  handleChangeFilter: (filter) => (dispatch({ type: CHANGE_FILTER, filter })),
-  handleVisibility: (filter) => (dispatch({ type: CHANGE_VISIBILITY, filter })),
-});
+const mapDispatchToProps = (dispatch) => ({ delete: (book) => dispatch(deleteBook(book)) });
 
 export default connect(mapStateToProps, mapDispatchToProps)(BookList);
