+++ src/index.css
@@ -30,7 +30,3 @@ code {
 #root {
   padding: 20px;
 }
-
-.form-label {
-  margin-left: 110px;
-}
