+++ src/stylesheets/NavBar.css
@@ -3,10 +3,17 @@
 }
 
 .login {
-  margin: 0;
-  padding: 14px;
+  width: 50px;
+  height: 50px;
+  padding: 4px;
   border: solid 1px #e8e8e8;
-  border-radius: 30px;
+  border-radius: 50%;
+}
+
+.login .fa {
+  font-size: 30px;
+  margin-left: 5px;
+  margin-top: 5px;
 }
 
 nav a {
@@ -14,9 +21,14 @@ nav a {
   font-family: 'Montserrat', sans-serif;
 }
 
+h1.title {
+  margin-top: 6px;
+  padding: 0;
+  width: 65%;
+}
+
 h1.title a {
   height: 37px;
-  margin: 6px 47px 2px 0;
   font-size: 30px;
   font-weight: bold;
   font-stretch: normal;
@@ -26,7 +38,7 @@ h1.title a {
   color: #0290ff;
 }
 
-h2.books,
+h2.books a,
 h2.categories a {
   color: var(--black-two);
 }
@@ -44,7 +56,7 @@ h2.books a {
 
 h2.categories {
   height: 16px;
-  margin: 18px 64px 11px 41px;
+  margin-top: 18px;
   opacity: 0.5;
   font-size: 13px;
   font-weight: normal;
@@ -52,6 +64,7 @@ h2.categories {
   font-style: normal;
   line-height: normal;
   letter-spacing: 1.9px;
+  width: 300px;
 }
 
 nav {
@@ -59,6 +72,16 @@ nav {
   justify-content: space-between;
   height: 95px;
   margin: 0 0 38px;
-  padding: 23px 99px 27px 100px;
+  padding: 23px 99px 27px 30px;
   background-color: #fff;
+}
+
+.nav-links {
+  display: flex;
+  justify-content: space-between;
+}
+
+.form-select {
+  position: relative;
+  top: -15px;
 }
